Normalize meta context before unifying.
The unification code expects that all DEFNs are fully substituted into
all problems.  However, in the spire code that calls the unifier, we
don't substitute into equations.  There are many ways to solve this
problem, but the easiest I could come up with was to substitute here,
since it's already well supported (the delayed subs).