Show that second bug only happens with derived substitution.
This was suspected from the analogies between lambdas and
unification-forall quantifiers, but the 'D1' and 'D2' data types make
this more clear.  The manual 'Subst' instance uses 'unsafeUnbind', but
there are no name collisions in the examples ('eg9D1', 'eg9D2', etc),
so capture should not be a concern.  Moreover, the problem also
happens in the "safe" code, and porting these examples should preserve
the problem ...